docs: 更新README文案和pom description refactor: 将启动类移动至module package下，减少扫包路径匹配消耗 pref: 为log.debug打印增加isDebugEnabled判断
