+++ bootstrap/src/main/java/cc/uncarbon/module/config/AsyncConfiguration.java
@@ -1,4 +1,4 @@
-package cc.uncarbon.config;
+package cc.uncarbon.module.config;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
