+++ bootstrap/src/main/java/cc/uncarbon/module/config/CustomInterceptorConfiguration.java
@@ -1,8 +1,8 @@
-package cc.uncarbon.config;
+package cc.uncarbon.module.config;
 
 import cc.uncarbon.framework.core.props.HelioProperties;
-import cc.uncarbon.interceptor.AdminSaTokenParseInterceptor;
-import cc.uncarbon.interceptor.DefaultSaTokenParseInterceptor;
+import cc.uncarbon.module.interceptor.AdminSaTokenParseInterceptor;
+import cc.uncarbon.module.interceptor.DefaultSaTokenParseInterceptor;
 import cc.uncarbon.module.adminapi.constant.AdminApiConstant;
 import cc.uncarbon.module.appapi.constant.AppApiConstant;
 import cc.uncarbon.module.sys.constant.SysConstant;
